fix example patch
