+++ internal/pkg/override/patch.go
@@ -5,6 +5,7 @@ package override
 
 import (
 	"fmt"
+	"path/filepath"
 	"strconv"
 	"strings"
 
@@ -13,6 +14,18 @@ import (
 )
 
 const jsonPointerSeparator = "/"
+
+func ScaffoldWithYAMLPatch(fs afero.Fs, dir string) error {
+	// If the directory does not exist, [afero.IsEmpty] returns false and an error.
+	// Therefore, we only want to check if a directory is empty only if it also exists.
+	exists, _ := afero.Exists(fs, dir)
+	isEmpty, _ := afero.IsEmpty(fs, dir)
+	if exists && !isEmpty {
+		return fmt.Errorf("directory %q is not empty", dir)
+	}
+
+	return templates.WalkOverridesYAMLPatchDir(writeFilesToDir(dir, fs))
+}
 
 // Patch applies overrides configured as JSON Patches,
 // as defined in https://www.rfc-editor.org/rfc/rfc6902.
@@ -53,7 +66,8 @@ func (p *Patch) Override(body []byte) ([]byte, error) {
 		return nil, fmt.Errorf("invalid template: %w", err)
 	}
 
-	for i, patch := range patches {
+	for i := range patches {
+		patch := patches[i] // needed because operations use pointer to patch.Value
 		var err error
 		switch patch.Operation {
 		case "add":
@@ -78,6 +92,7 @@ func (p *Patch) Override(body []byte) ([]byte, error) {
 }
 
 func unmarshalPatches(path string, fs afero.Fs) ([]yamlPatch, error) {
+	path = filepath.Join(path, "cfn.patches.yml")
 	content, err := afero.ReadFile(fs, path)
 	if err != nil {
 		return nil, fmt.Errorf("read file at %q: %w", path, err)
